+++ packaging/debian/validate-package-assets.sh
@@ -66,6 +66,7 @@ require_text "$sysusers" "g dasobjectstore"
 
 require_text "$tmpfiles" "z /srv/dasobjectstore 0750 dasobjectstore dasobjectstore -"
 require_text "$tmpfiles" "d /opt/dasobjectstore 0750 dasobjectstore dasobjectstore -"
+require_text "$tmpfiles" "d /var/lib/dasobjectstore/object-service 0700 dasobjectstore dasobjectstore -"
 require_text "$tmpfiles" "d /var/lib/dasobjectstore/report-rebuild 0750 dasobjectstore dasobjectstore -"
 require_text "$tmpfiles" "d /var/lib/dasobjectstore/telemetry 0750 dasobjectstore dasobjectstore -"
 require_text "$web_config" "\"bind_address\": \"0.0.0.0\""
@@ -78,6 +79,7 @@ require_text "$postinst" "service_user=\"dasobjectstore\""
 require_text "$postinst" "managed_root=\"/srv/dasobjectstore\""
 require_text "$postinst" "product_root=\"/opt/dasobjectstore\""
 require_text "$postinst" 'ensure_owned_dir /var/lib/dasobjectstore/report-rebuild 0750'
+require_text "$postinst" 'ensure_owned_dir /var/lib/dasobjectstore/object-service 0700'
 require_text "$postinst" 'ensure_owned_dir /var/lib/dasobjectstore/telemetry 0750'
 require_text "$postinst" 'ensure_container_runtime_access'
 require_text "$postinst" 'usermod -aG docker "$service_user"'
@@ -139,6 +141,7 @@ require_text "$build_rpm" 'usr/lib/tmpfiles.d/dasobjectstore.conf'
 require_text "$build_rpm" 'systemd-sysusers /usr/lib/sysusers.d/dasobjectstore.conf'
 require_text "$build_rpm" 'systemd-tmpfiles --create /usr/lib/tmpfiles.d/dasobjectstore.conf'
 require_text "$build_rpm" 'install -d -o "\$service_user" -g "\$service_group" -m 0750 /var/lib/dasobjectstore/report-rebuild'
+require_text "$build_rpm" 'install -d -o "\$service_user" -g "\$service_group" -m 0700 /var/lib/dasobjectstore/object-service'
 require_text "$build_rpm" 'install -d -o "\$service_user" -g "\$service_group" -m 0750 /var/lib/dasobjectstore/telemetry'
 require_text "$build_rpm" '/usr/libexec/dasobjectstore/dasobjectstore-local-auth-helper'
 require_text "$build_rpm" '/usr/libexec/dasobjectstore/gnostikon-workflow-control'