+++ packaging/rpm/build-rpm.sh
@@ -160,6 +160,7 @@ fi
 
 install -d -o "\$service_user" -g "\$service_group" -m 0750 /run/dasobjectstore
 install -d -o "\$service_user" -g "\$service_group" -m 0750 /var/lib/dasobjectstore
+install -d -o "\$service_user" -g "\$service_group" -m 0700 /var/lib/dasobjectstore/object-service
 install -d -o "\$service_user" -g "\$service_group" -m 0750 /var/lib/dasobjectstore/report-rebuild
 install -d -o "\$service_user" -g "\$service_group" -m 0750 /var/lib/dasobjectstore/telemetry
 install -d -o "\$service_user" -g "\$service_group" -m 0750 /var/log/dasobjectstore